Movement is broken
After fixing all the merges our key input is broken for both the jump and crouch. I am trying to debut and check settings to see if it's a coding error or a project setttings thing we need to adjust.
Maybe look into placing debug console prints to determine if key inputs are working first before altering any more code.