welp...
I attmepted to recreate the code and not  I got a lot of errros and stuff. I am trying to refactor into a gamemanger but this is taking a long time.

I will look further into this anotehr time
If you want go ahead and just review what I did

ALSO SORRY
I was messing with the onclick events for the main menu and broke some things. I tried to fix but I am too tired for this now
I will look more into this tomorrow