Code updates
In this amazing update I was able to make somethings and only break a small bit of my code.
we officailly have a obstacle spawner, but I made it so the obts go directy towarsd the palyer.... aka if player jump it follow

I will need to create a PLAYER starting pos and have obs go to playerStartingPos-20 it go behind then despawn once it enters the bubble.
I will be doing another update in a few hours

I also put things into folders for us. Just so it's a bit easier to find the scrips we are working on