End Game, Store Manager, scene cleanups
So far I created the sytem to end the game. Store manager is currently in the making I will finish this in the monring when I wake up.
I am currently working on the UI for the store. this will be hopefully done tomorrow andI will work on importing to the iphone as well